+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.4-unofficial-mushymato.1](https://github.com/Mushymato/StardewUI/releases/tag/0.6.4-unofficial-mushymato.1)
+
+### Changed
+- `DropDownList` search feature is now locked behind property `Searchable`.
+
 ## [0.6.4-unofficial-mushymato.0](https://github.com/Mushymato/StardewUI/releases/tag/0.6.4-unofficial-mushymato.0)
 
 ### Added
@@ -34,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   - The two inner label views (text and placeholder) no longer scales with the outer view.
 - `PositionSelector` will now also set the initial `xPositionOnScreen` and `yPositionOnScreen` values.
 - `Expander` now uses a caret with square source rect to avoid stretching when rotated.
+- `DropDownList` can now accept key inputs and be searched.
 
 ### Fixed
 - Segments not measuring initial layout properly.